mysql: first shot at implementing refdb lookup
Performs the operation in the expected manner: executes a select query
selecting the oid of any reference that matches the desired reference
name. Extracts the OID back out, creates a git reference around it.

This does not allow for setting symbolic git references -- too bad.

Additionally, the world ends if more than one row is returned for the
desired reference. This can only happen if the db storing these
references is malformed or corrupt in some manner, so there's not a lot
of scope for nice error handling here.