mysql: implement refdb exists via lookup method
Nothing in libgit2 appears to use 'exists', neither is it an exported
API (apparently). In particular, there's nothing to indicate what the
return code should be in the "exists" parameter. In lieu of that, make a
guess and carry on. The refdb_backend header indicates that this member
is mandatory