mysql: implement reference writing facility
One more there's no particularly interesting logic here: the reference
is looked up to see whether or not it exists yet (an error is returned
if it does and the force flag is not specified). The existing reference
is deleted if it existed. After this, an insert is performed to create
the reference in the database.