mysql: Pull in data for reference iteration
This sucks all the matched references and OIDs into memory, stored in
the iterator structure. This has various bad side-effects if the number
of references matched is large: this is a consideration for the user,
if they really have that many references.

I've done it this way becdause, as I understand it, the mysql db
connection will be busy while there are still rows to be read from a
result set.  If we didn't pull all the rows out immediately, one would
only be able to use one reference iterator at a time, which is an
unreasonable limitation.