mysql: support 'writepack' function (code bomb)
libgit2 does not expose any useful facilities for parsing raw packfiles.
This makes it almost impossible for the mysql backend to simply suck a
packfile into memory and insert stuff into the database. So instead,
write the packfile out to a dir in /tmp, using the published 'indexer'
facility. (It's opaque, so we can't pull indexed objects out). Then open
that packfile with existing odb facilities, to be able to access
objects.

Once this has been done, we can start putting information into the mysql
db. This must be atomic, however. To avoid either locking tables, long
transactions, or having to manually unwind, instead insert all objects
into a temporary table. The contents thereof are then inserted into the
real table, atomically, once all objects are in the temp one.