mysql: support symbolic reference names
This works by storing the reference type and (potentially null) symbolic
name in the database. If either the OID or symbolic ref name are null in
the source reference, they're null in the database, and vice versa.

I wasn't intending on implementing this, however it seems a large number
of things immediately croak if there's no HEAD reference, which almost
certainly needs to be symbolic.